+++ src/generator/template-ops/TemplateVariables.model.ts
@@ -3,10 +3,13 @@ export interface TemplateVariables {
   componentPrefix: string;
   /** Ex: my-component */
   dashCaseName: string;
+  /** Ex: myComponent */
+  camelCaseName: string;
   /** Ex: MyComponent */
   upperCamelCaseName: string;
   /** Ex: MY_COMPONENT */
   constantCaseName: string;
   /** Ex: My Component */
   upperReadableName: string;
+  type: string;
 }